// libs/_common/availability/availability.enum.ts
export enum AvailabilityConstants {
  REAL_DAY_MINUTES = 1440,
  REAL_FIRST_DAY_MINUTE = 0,
  REAL_LAST_DAY_MINUTE = 1440,
  FIRST_POSSIBLE_DAY_MINUTE_BASED_ON_TIMEZONES = -1440,
  LAST_POSSIBLE_DAY_MINUTE_BASED_ON_TIMEZONES = 2880,
  FIRST_DAY_IN_WEEK_INDEX = 0,
  LAST_DAY_IN_WEEK_INDEX = 6,
  WEEK_DAYS_LENGTH = 7,
  SUNDAY_INDEX_AT_WEEK = 0,
  SATURDAY_INDEX_AT_WEEK = 6
}

export enum AvailabilityAppointmentStatusEnum {
  SUCCEEDED = 'SUCCEEDED',
  IN_PROGRESS = 'IN_PROGRESS',
  REQUEST_SENT_TO_DOCTOR = 'REQUEST_SENT_TO_DOCTOR',
  DOCTOR_ACCEPTED_REQUEST = 'DOCTOR_ACCEPTED_REQUEST'
}